Move classified track density computation to _track_density.py
_density_weighting.py keeps the spec dataclasses and the small
spec-reading tasks; the trajectory-specific composite task and its
branch helpers now live in _track_density.py with a one-way import
of the specs.
